Replace float variables with double variables
# Conflicts:
#	simplex/simpl.cpp
#	simplex/simpl.h
#	simplex/simplexTableau.h

+++ simplex/simpl.cpp
@@ -16,7 +16,7 @@ simpl::~simpl() {
 	destroyTab();
 }
 
-bool simpl::changeValue(float value, int row, int col) {
+bool simpl::changeValue(double value, int row, int col) {
 	if (canChangeValue)
 	{
 		tableauErrorCode errorCode = tableau->changeValue(value, row, col);
@@ -38,7 +38,7 @@ bool simpl::pivot(int row, int col) {
 	return (errorCode == TABLEAU_SUCCESS);
 }
 
-simplexErrorCode simpl::simplex(float* xValuePtr, float* optimalValue) {
+simplexErrorCode simpl::simplex(double* xValuePtr, double* optimalValue) {
 	bool simplexdone = false;
 	int i;
 	int colnum;
